Caching relative velocity

+++ Apps/BASimulator/Problems/HairyBallWeta.cc
@@ -586,7 +586,7 @@ void HairyBallWeta::serialize( std::ofstream& of )
   TopologicalObjectSerializer objserializer;
 
   // Serialize the number of triangle meshes
-  int numtrimeshes = m_tri_meshes.size();
+  int numtrimeshes = (int) m_tri_meshes.size();
   assert( numtrimeshes == 1 );
   serializeVal(of,numtrimeshes);
 
@@ -598,7 +598,7 @@ void HairyBallWeta::serialize( std::ofstream& of )
   }
 
   // Serialize the number of rods
-  int numrods = m_rods.size();
+  int numrods = (int) m_rods.size();
   serializeVal(of,numrods);  
 
   // Serialize each rod
